requirements.txt
attrs==21.4.0
Authlib==0.15.5
Cerberus==1.3.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
coverage==6.3.2
cryptography==36.0.1
Flask==2.0.3
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
pytest-cov==3.0.0
requests==2.27.1
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.8
URLObject==2.4.3
Werkzeug==2.0.3
